+++ src/RabbitLink.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 15.0.26730.10
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RabbitLink", "RabbitLink\RabbitLink.csproj", "{C2948199-17DA-4847-B203-2B4E6F96BAAF}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RabbitLink.Tests", "RabbitLink.Tests\RabbitLink.Tests.csproj", "{7A537FC1-7715-4DAC-988C-5E9982DCF621}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RabbitLink.UnitTests", "RabbitLink.UnitTests\RabbitLink.UnitTests.csproj", "{7A537FC1-7715-4DAC-988C-5E9982DCF621}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Playground", "Playground\Playground.csproj", "{9CFDFF2F-8716-42AD-B780-31F9CB6AE42C}"
 EndProject
@@ -20,6 +20,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RabbitLink.Serialization.Json", "RabbitLink.Serialization.Json\RabbitLink.Serialization.Json.csproj", "{6A5A7A43-295E-4DF2-B4FB-0858B2343A28}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RabbitLink.FunctionalTests", "RabbitLink.FunctionalTests\RabbitLink.FunctionalTests.csproj", "{31954DE5-8DAE-4454-892B-81056E9B0E71}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,6 +45,10 @@ Global
 		{6A5A7A43-295E-4DF2-B4FB-0858B2343A28}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6A5A7A43-295E-4DF2-B4FB-0858B2343A28}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6A5A7A43-295E-4DF2-B4FB-0858B2343A28}.Release|Any CPU.Build.0 = Release|Any CPU
+		{31954DE5-8DAE-4454-892B-81056E9B0E71}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{31954DE5-8DAE-4454-892B-81056E9B0E71}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{31954DE5-8DAE-4454-892B-81056E9B0E71}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{31954DE5-8DAE-4454-892B-81056E9B0E71}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
